the kernel needs to accept uint32_t IRQ numbers, like exec, so that we can provide more than the base "256" device IRQS's (e.g, for MSI).
git-svn-id: https://svn.aros.org/svn/aros/trunk@53993 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/kernel/addirqhandler.c
@@ -26,7 +26,7 @@
         AROS_LH4(void *, KrnAddIRQHandler,
 
 /*  SYNOPSIS */
-        AROS_LHA(uint8_t, irq, D0),
+        AROS_LHA(uint32_t, irq, D0),
         AROS_LHA(irqhandler_t *, handler, A0),
         AROS_LHA(void *, handlerData, A1),
         AROS_LHA(void *, handlerData2, A2),
